mgr/dashboard: Move validators into forms directory
Now that 'forms' directory exists, the validators are moved to
it. Angular also does this with it's validators ('@angular/forms').
